+++ src/executor.rs
@@ -1,0 +1,110 @@
+use std::mem::replace;
+use pyo3::PyResult;
+use revm::{Context, Evm, EvmContext, FrameOrResult, FrameResult};
+use revm::primitives::{ExecutionResult};
+use revm::primitives::TransactTo;
+use revm_interpreter::{CallInputs, CreateInputs, SuccessOrHalt};
+use crate::database::DB;
+use crate::utils::pyerr;
+
+pub(crate) fn evm_call<EXT>(mut evm: Evm<'_, EXT, DB>) -> PyResult<(ExecutionResult, EvmContext<DB>)> {
+    evm.handler.validation().env(&evm.context.evm.env).map_err(pyerr)?;
+    let initial_gas_spend = evm
+        .handler
+        .validation()
+        .initial_tx_gas(&evm.context.evm.env)
+        .map_err(pyerr)?;
+    evm.handler
+        .validation()
+        .tx_against_state(&mut evm.context)
+        .map_err(pyerr)?;
+
+    let ctx = &mut evm.context;
+    let pre_exec = evm.handler.pre_execution();
+
+    // load access list and beneficiary if needed.
+    pre_exec.load_accounts(ctx).map_err(pyerr)?;
+
+    // load precompiles
+    ctx.evm.set_precompiles(pre_exec.load_precompiles());
+
+    // deduce caller balance with its limit.
+    pre_exec.deduct_caller(ctx).map_err(pyerr)?;
+
+    let gas_limit = ctx.evm.env.tx.gas_limit - initial_gas_spend;
+
+    let exec = evm.handler.execution();
+    // call inner handling of call/create
+    let first_frame_or_result = match ctx.evm.env.tx.transact_to {
+        TransactTo::Call(_) => exec.call(
+            ctx,
+            CallInputs::new_boxed(&ctx.evm.env.tx, gas_limit).unwrap(),
+        ).map_err(pyerr)?,
+        TransactTo::Create(_) => exec.create(
+            ctx,
+            CreateInputs::new_boxed(&ctx.evm.env.tx, gas_limit).unwrap(),
+        ).map_err(pyerr)?,
+    };
+
+    // Starts the main running loop.
+    let mut result = match first_frame_or_result {
+        FrameOrResult::Frame(first_frame) => evm.start_the_loop(first_frame).map_err(pyerr)?,
+        FrameOrResult::Result(result) => result,
+    };
+
+    let ctx = &mut evm.context;
+
+    // handle output of call/create calls.
+    evm.handler
+        .execution()
+        .last_frame_return(ctx, &mut result)
+        .map_err(pyerr)?;
+
+    let post_exec = evm.handler.post_execution();
+    // Reimburse the caller
+    post_exec.reimburse_caller(ctx, result.gas()).map_err(pyerr)?;
+    // Reward beneficiary
+    post_exec.reward_beneficiary(ctx, result.gas()).map_err(pyerr)?;
+    // Returns output of transaction.
+    Ok((output(ctx, result)?, evm.context.evm))
+}
+
+/// Main return handle, returns the output of the transaction.
+#[inline]
+fn output<EXT>(
+    context: &mut Context<EXT, DB>,
+    result: FrameResult,
+) -> PyResult<ExecutionResult> {
+    replace(&mut context.evm.error, Ok(())).map_err(pyerr)?;
+    // used gas with refund calculated.
+    let gas_refunded = result.gas().refunded() as u64;
+    let final_gas_used = result.gas().spend() - gas_refunded;
+    let output = result.output();
+    let instruction_result = result.into_interpreter_result();
+
+    let result = match instruction_result.result.into() {
+        SuccessOrHalt::Success(reason) => ExecutionResult::Success {
+            reason,
+            gas_used: final_gas_used,
+            gas_refunded,
+            logs: vec![], // todo: logs
+            output,
+        },
+        SuccessOrHalt::Revert => ExecutionResult::Revert {
+            gas_used: final_gas_used,
+            output: output.into_data(),
+        },
+        SuccessOrHalt::Halt(reason) => ExecutionResult::Halt {
+            reason,
+            gas_used: final_gas_used,
+        },
+        // Only two internal return flags.
+        SuccessOrHalt::FatalExternalError
+        | SuccessOrHalt::InternalContinue
+        | SuccessOrHalt::InternalCallOrCreate => {
+            panic!("Internal return flags should remain internal {instruction_result:?}")
+        }
+    };
+
+    Ok(result)
+}
