+++ src/utils.rs
@@ -1,10 +1,8 @@
 use pyo3::{exceptions::PyTypeError, prelude::*};
-use revm::primitives::{Address, State, HashMap as RevmHashMap};
+use revm::primitives::{Address, HashMap as RevmHashMap};
 use std::fmt::Debug;
 use pyo3::exceptions::PyRuntimeError;
-use std::collections::HashMap;
 use revm::db::DbAccount;
-use crate::AccountInfo;
 use crate::types::PyDB;
 
 pub(crate) fn addr(s: &str) -> Result<Address, PyErr> {
@@ -17,12 +15,6 @@ pub(crate) fn addr_or_zero(s: Option<&str>) -> Result<Address, PyErr> {
         Some(s) => addr(s),
         None => Ok(Address::ZERO),
     }
-}
-
-pub(crate) fn pydict(res: State) -> HashMap<String, AccountInfo> {
-    res.iter().map(
-        |(address, db_acc)| (address.to_string(), db_acc.info.clone().into())
-    ).collect()
 }
 
 /// Convert a Rust error into a Python error.
